Added soul objects for each Soul. Added custom_signal to player
Made the Soul objects and have them extending Soul. Saw in a comment in the player script that an "upgrade_gained" signal is supposed to connect to the unlock_upgrade function, so I created the custom signal and made the connection in _ready() but I'm not sure what else to do with it since the Soul template is already calling the unlock_upgrade function.